LayoutTests/fast/text/font-stretch-parse.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div style="font: 16px/18px 'Times';">
<div id="test1"></div>
<div id="test2" style="font-stretch: 1%;"></div>
<div id="test3" style="font-stretch: 2;"></div>
<div id="test4" style="font-stretch: ultra-condensed;"></div>
<div id="test5" style="font-stretch: extra-condensed;"></div>
<div id="test6" style="font-stretch: condensed;"></div>
<div id="test7" style="font-stretch: semi-condensed;"></div>
<div id="test8" style="font-stretch: normal;"></div>
<div id="test9" style="font-stretch: semi-expanded;"></div>
<div id="test10" style="font-stretch: expanded;"></div>
<div id="test11" style="font-stretch: extra-expanded;"></div>
<div id="test12" style="font-stretch: ultra-expanded;"></div>
<div id="test13" style="font-stretch: garbage;"></div>
<div id="test14" style="font-stretch: initial;"></div>
<div id="test15" style="font-stretch: calc(3 + 4);"></div>
<div id="test16" style="font: 100 extra-condensed 48px/49px 'Helvetica Neue';"></div>
<div id="test17" style="font: extra-condensed 100 48px/49px 'Helvetica Neue';"></div>
<div id="test18" style="font: 100 48px/49px 'Helvetica Neue';"></div>
<div style="font-stretch: extra-expanded;"><div id="test19" style="font: 100 48px/49px 'Helvetica Neue';"></div></div>
<div id="test20" style="font: italic small-caps 100 extra-expanded 48px/49px 'Helvetica Neue';"></div>
<div id="test21" style="font-stretch: calc(3% + 4%);"></div>
</div>
<script>
shouldBeEqualToString("getComputedStyle(document.getElementById('test1')).fontStretch", "100%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test2')).fontStretch", "1%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test3')).fontStretch", "100%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test4')).fontStretch", "50%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test5')).fontStretch", "62.5%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test6')).fontStretch", "75%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test7')).fontStretch", "87.5%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test8')).fontStretch", "100%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test9')).fontStretch", "112.5%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test10')).fontStretch", "125%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test11')).fontStretch", "150%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test12')).fontStretch", "200%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test13')).fontStretch", "100%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test14')).fontStretch", "100%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test15')).fontStretch", "100%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test16')).fontStretch", "62.5%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test17')).fontStretch", "62.5%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test18')).fontStretch", "100%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test19')).fontStretch", "100%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test20')).fontStretch", "150%");
shouldBeEqualToString("getComputedStyle(document.getElementById('test21')).fontStretch", "7%");

shouldBeEqualToString("getComputedStyle(document.getElementById('test1')).font", "16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test2')).font", "");
shouldBeEqualToString("getComputedStyle(document.getElementById('test3')).font", "16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test4')).font", "ultra-condensed 16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test5')).font", "extra-condensed 16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test6')).font", "condensed 16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test7')).font", "semi-condensed 16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test8')).font", "16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test9')).font", "semi-expanded 16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test10')).font", "expanded 16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test11')).font", "extra-expanded 16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test12')).font", "ultra-expanded 16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test13')).font", "16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test14')).font", "16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test15')).font", "16px / 18px Times");
shouldBeEqualToString("getComputedStyle(document.getElementById('test16')).font", `100 extra-condensed 48px / 49px "Helvetica Neue"`);
shouldBeEqualToString("getComputedStyle(document.getElementById('test17')).font", `100 extra-condensed 48px / 49px "Helvetica Neue"`);
shouldBeEqualToString("getComputedStyle(document.getElementById('test18')).font", `100 48px / 49px "Helvetica Neue"`);
shouldBeEqualToString("getComputedStyle(document.getElementById('test19')).font", `100 48px / 49px "Helvetica Neue"`);
shouldBeEqualToString("getComputedStyle(document.getElementById('test20')).font", `italic small-caps 100 extra-expanded 48px / 49px "Helvetica Neue"`);
shouldBeEqualToString("getComputedStyle(document.getElementById('test21')).font", "");

shouldBeEqualToString("document.getElementById('test1').style.font", "");
shouldBeEqualToString("document.getElementById('test16').style.font", `100 extra-condensed 48px / 49px "Helvetica Neue"`);
shouldBeEqualToString("document.getElementById('test17').style.font", `100 extra-condensed 48px / 49px "Helvetica Neue"`);
shouldBeEqualToString("document.getElementById('test18').style.font", `100 48px / 49px "Helvetica Neue"`);
shouldBeEqualToString("document.getElementById('test19').style.font", `100 48px / 49px "Helvetica Neue"`);
shouldBeEqualToString("document.getElementById('test20').style.font", `italic small-caps 100 extra-expanded 48px / 49px "Helvetica Neue"`);
</script>
</body>
</html>
