LayoutTests/fast/text/font-stretch-parse-expected.txt

PASS getComputedStyle(document.getElementById('test1')).fontStretch is "100%"
PASS getComputedStyle(document.getElementById('test2')).fontStretch is "1%"
PASS getComputedStyle(document.getElementById('test3')).fontStretch is "100%"
PASS getComputedStyle(document.getElementById('test4')).fontStretch is "50%"
PASS getComputedStyle(document.getElementById('test5')).fontStretch is "62.5%"
PASS getComputedStyle(document.getElementById('test6')).fontStretch is "75%"
PASS getComputedStyle(document.getElementById('test7')).fontStretch is "87.5%"
PASS getComputedStyle(document.getElementById('test8')).fontStretch is "100%"
PASS getComputedStyle(document.getElementById('test9')).fontStretch is "112.5%"
PASS getComputedStyle(document.getElementById('test10')).fontStretch is "125%"
PASS getComputedStyle(document.getElementById('test11')).fontStretch is "150%"
PASS getComputedStyle(document.getElementById('test12')).fontStretch is "200%"
PASS getComputedStyle(document.getElementById('test13')).fontStretch is "100%"
PASS getComputedStyle(document.getElementById('test14')).fontStretch is "100%"
PASS getComputedStyle(document.getElementById('test15')).fontStretch is "100%"
PASS getComputedStyle(document.getElementById('test16')).fontStretch is "62.5%"
PASS getComputedStyle(document.getElementById('test17')).fontStretch is "62.5%"
PASS getComputedStyle(document.getElementById('test18')).fontStretch is "100%"
PASS getComputedStyle(document.getElementById('test19')).fontStretch is "100%"
PASS getComputedStyle(document.getElementById('test20')).fontStretch is "150%"
PASS getComputedStyle(document.getElementById('test21')).fontStretch is "7%"
PASS getComputedStyle(document.getElementById('test1')).font is "16px / 18px Times"
PASS getComputedStyle(document.getElementById('test2')).font is "16px / 18px Times"
PASS getComputedStyle(document.getElementById('test3')).font is "16px / 18px Times"
PASS getComputedStyle(document.getElementById('test4')).font is "ultra-condensed 16px / 18px Times"
PASS getComputedStyle(document.getElementById('test5')).font is "extra-condensed 16px / 18px Times"
PASS getComputedStyle(document.getElementById('test6')).font is "condensed 16px / 18px Times"
PASS getComputedStyle(document.getElementById('test7')).font is "semi-condensed 16px / 18px Times"
PASS getComputedStyle(document.getElementById('test8')).font is "16px / 18px Times"
PASS getComputedStyle(document.getElementById('test9')).font is "semi-expanded 16px / 18px Times"
PASS getComputedStyle(document.getElementById('test10')).font is "expanded 16px / 18px Times"
PASS getComputedStyle(document.getElementById('test11')).font is "extra-expanded 16px / 18px Times"
PASS getComputedStyle(document.getElementById('test12')).font is "ultra-expanded 16px / 18px Times"
PASS getComputedStyle(document.getElementById('test13')).font is "16px / 18px Times"
PASS getComputedStyle(document.getElementById('test14')).font is "16px / 18px Times"
PASS getComputedStyle(document.getElementById('test15')).font is "16px / 18px Times"
PASS getComputedStyle(document.getElementById('test16')).font is "100 extra-condensed 48px / 49px \"Helvetica Neue\""
PASS getComputedStyle(document.getElementById('test17')).font is "100 extra-condensed 48px / 49px \"Helvetica Neue\""
PASS getComputedStyle(document.getElementById('test18')).font is "100 48px / 49px \"Helvetica Neue\""
PASS getComputedStyle(document.getElementById('test19')).font is "100 48px / 49px \"Helvetica Neue\""
PASS getComputedStyle(document.getElementById('test20')).font is "italic small-caps 100 extra-expanded 48px / 49px \"Helvetica Neue\""
PASS getComputedStyle(document.getElementById('test21')).font is "16px / 18px Times"
PASS document.getElementById('test1').style.font is ""
PASS document.getElementById('test16').style.font is "100 extra-condensed 48px / 49px \"Helvetica Neue\""
PASS document.getElementById('test17').style.font is "100 extra-condensed 48px / 49px \"Helvetica Neue\""
PASS document.getElementById('test18').style.font is "100 48px / 49px \"Helvetica Neue\""
PASS document.getElementById('test19').style.font is "100 48px / 49px \"Helvetica Neue\""
PASS document.getElementById('test20').style.font is "italic small-caps 100 extra-expanded 48px / 49px \"Helvetica Neue\""
PASS successfullyParsed is true

TEST COMPLETE
